chore: Rename to `settingsFragment`
