+++ lib/ui/views/settings/settingsFragment/settings_manage_sources.dart
