+++ lib/ui/views/settings/settings_view.dart
@@ -3,8 +3,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_i18n/flutter_i18n.dart';
 import 'package:google_fonts/google_fonts.dart';
-import 'package:revanced_manager/ui/views/settings/settingsFragement/settings_update_language.dart';
-import 'package:revanced_manager/ui/views/settings/settingsFragement/settings_update_theme.dart';
+import 'package:revanced_manager/ui/views/settings/settingsFragment/settings_update_language.dart';
+import 'package:revanced_manager/ui/views/settings/settingsFragment/settings_update_theme.dart';
 import 'package:revanced_manager/ui/views/settings/settings_viewmodel.dart';
 import 'package:revanced_manager/ui/widgets/settingsView/settings_advanced_section.dart';
 import 'package:revanced_manager/ui/widgets/settingsView/settings_export_section.dart';
