+++ lib/ui/views/settings/settings_viewmodel.dart
@@ -9,8 +9,8 @@ import 'package:revanced_manager/app/app.router.dart';
 import 'package:revanced_manager/services/manager_api.dart';
 import 'package:revanced_manager/services/toast.dart';
 import 'package:revanced_manager/ui/views/patcher/patcher_viewmodel.dart';
-import 'package:revanced_manager/ui/views/settings/settingsFragement/settings_update_language.dart';
-import 'package:revanced_manager/ui/views/settings/settingsFragement/settings_update_theme.dart';
+import 'package:revanced_manager/ui/views/settings/settingsFragment/settings_update_language.dart';
+import 'package:revanced_manager/ui/views/settings/settingsFragment/settings_update_theme.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:share_extend/share_extend.dart';
 import 'package:stacked/stacked.dart';
